orgorush-hackathon/backend/orgo-payment-backend/requirements.txt
flask==2.3.3
fastapi==0.104.1
uvicorn==0.24.0
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO

solana==0.32.0
anchorpy==0.19.1
cryptography==41.0.7

numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2

py_ecc==6.0.0

sqlalchemy==2.0.23
requests==2.31.0
httpx<0.24.0,>=0.23.0
aiohttp==3.9.1

python-dotenv==1.0.0
pydantic==2.5.0

bcrypt==4.1.2
pyjwt==2.8.0

pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

click==8.1.7
python-multipart==0.0.6
websockets<12.0,>=9.0


